requirements.txt
gensim==4.2.0
gplearn==0.4.2
h5py
ipykernel==6.16.2
ipython==7.34.0
jieba==0.42.1
lightgbm==3.3.5
matplotlib==3.5.3
numpy
pandas==1.3.5
Pillow==9.5.
pytorch-tabnet==4.0
pytz==2023.3
scikit-learn==1.0.2
scipy
torch
torchaudio
torchvision
tqdm==4.65.0